requirements.txt
sentence_transformers
ollama
psycopg2-binary==2.9.7
pandas==2.1.4
numpy==1.24.3
